sync patch r10893 from main trunk. Add security check to make code run safe.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/branches/UDK2010@11552 6f19259b-4bc3-4df7-8a09-765794883524

+++ OptionRomPkg/UndiRuntimeDxe/Init.c
@@ -882,7 +882,7 @@ AppendMac2DevPath (
 
   MacAddrNode.Header.Type       = MESSAGING_DEVICE_PATH;
   MacAddrNode.Header.SubType    = MSG_MAC_ADDR_DP;
-  MacAddrNode.Header.Length[0]  = sizeof (MacAddrNode);
+  MacAddrNode.Header.Length[0]  = (UINT8) sizeof (MacAddrNode);
   MacAddrNode.Header.Length[1]  = 0;
 
   //
